requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
httpx>=0.27
beautifulsoup4>=4.12
openai>=1.50
python-multipart>=0.0.9
pyyaml>=6.0
itsdangerous>=2.1
lxml>=5.0
